fix: use unique unnamed service name value
In some places, a literal string was being used when a service
name is not provided via the configuration interface. This value
however differed from the DEFAULT_SERVICE_NAME constant, resulting
in multiple service names being used in different parts of the
library. This change ensures that only one default service name is
used across the library.