ci: redirect DI exploration output to file
The output from the DI exploration tools might be overwhelming and
make the debugging of issues during test runs a bit hard to grok.
We add the option of redirecting the output to a file that can be
rendered in a different step, to keep the test run output cleaner.