feat(propagation): use 128-bit trace ids by default (#7181)
This PR looks really intimidating, but it's not that bad, it's just that
we now have a new tag `_dd.p.tid` on almost all snapshots, so that
accounts for a vast majority of the files changed. It should be noted
that the _dd.p.tid value is not correct in most instances, and the
test-agent has been set to ignore the value of it, as it does with other
attributes e.g. runtime-id etc. We have other tests that cover making
sure the value is correct.


**Context:** 
The basic idea for 128-bit propagation for backwards compatibility is
that during injection, we still set `x-datadog-trace-id` header to the
lower 64-bits of the 128-bit id. The 64 highest order bits are encoded
in base 16 and stored in the `_dd.p.tid` in `x-datadog-tags` header. We
then combine them upon extraction to create the full 128-bit trace id.

**What this PR does:**
1. Enables 128-bit trace ids by default. Note: that 128-bit log
injection default is still False.
2. Adds some validation and logic around what we do when receiving
`_dd.p.tid` in the `x-datadog-tags` header.
3. Factors out some of the logic used in `_extract` to make it easier to
test.

**Testing**
The changes to the tests are almost all just to compare the new 128-bit
value properly to the combined value of the headers. Or to add some
logic depending on if 128-bit is enabled or not. The system-tests cover
the validation logic and when to add `_dd.p.tid`.


## Checklist

- [x] Change(s) are motivated and described in the PR description.
- [x] Testing strategy is described if automated tests are not included
in the PR.
- [x] Risk is outlined (performance impact, potential for breakage,
maintainability, etc).
- [x] Change is maintainable (easy to change, telemetry, documentation).
- [x] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
are followed. If no release note is required, add label
`changelog/no-changelog`.
- [x] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/)).
- [x] Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist

- [x] Title is accurate.
- [x] No unnecessary changes are introduced.
- [x] Description motivates each change.
- [x] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes unless absolutely necessary.
- [x] Testing strategy adequately addresses listed risk(s).
- [x] Change is maintainable (easy to change, telemetry, documentation).
- [x] Release note makes sense to a user of the library.
- [x] Reviewer has explicitly acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment.
- [x] Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)
- [x] If this PR touches code that signs or publishes builds or
packages, or handles credentials of any kind, I've requested a review
from `@DataDog/security-design-and-guidance`.
- [x] This PR doesn't touch any of that.