feat(propagation): assess and link conflicting contexts upon extraction (#7590)
**What?**
Libraries must add the ability to read multiple propagator formats so
that if the subsequent trace contexts with conflicting trace-ids are
detected then an appropriate span link will be added to the span
indicating the error.

**Design**
We need to add the span_links to the Context object, and can't go
directly to adding them to the span. Therefore the links are added to
the span on `start_span`, we add the span_links to the span if it's the
local root span.

**Testing**
For the tests I wanted to make sure the Context object looked full
correct so I created a test to look at the full object.
I did include a test in which a span link is added, and also the
tracecontext trace_id matches the primary context's trace_id so the the
tracestate is added to the context, which is the behavior implemented
[here](https://github.com/DataDog/dd-trace-py/pull/7538). However, since
this PR essentially adds to that functionality I thought it good to
include.

**Peformance**
Performance overhead of just adding SpanLink upon non-matching trace_ids
for contexts is pretty minimal, each of these cases would generate one
span link. If a user enabled all styles and had non-matching trace_ids
for 3 styles it would be worse, but that seems pretty unusual. This case
seems by far the most typical.
![config yaml — dd-trace-py 2023-11-16 at 12 20 27
AM](https://github.com/DataDog/dd-trace-py/assets/32471391/ee0e1224-3e5e-484a-ab75-24ab3f976a7a)




## Checklist

- [x] Change(s) are motivated and described in the PR description.
- [x] Testing strategy is described if automated tests are not included
in the PR.
- [x] Risk is outlined (performance impact, potential for breakage,
maintainability, etc).
- [x] Change is maintainable (easy to change, telemetry, documentation).
- [x] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
are followed. If no release note is required, add label
`changelog/no-changelog`.
- [x] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/)).
- [x] Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist

- [x] Title is accurate.
- [x] No unnecessary changes are introduced.
- [x] Description motivates each change.
- [x] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes unless absolutely necessary.
- [x] Testing strategy adequately addresses listed risk(s).
- [x] Change is maintainable (easy to change, telemetry, documentation).
- [x] Release note makes sense to a user of the library.
- [x] Reviewer has explicitly acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment.
- [x] Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)
- [x] If this PR touches code that signs or publishes builds or
packages, or handles credentials of any kind, I've requested a review
from `@DataDog/security-design-and-guidance`.
- [x] This PR doesn't touch any of that.

---------

Co-authored-by: Emmett Butler <723615+emmettbutler@users.noreply.github.com>