fix(bootstrap): do not load new modules after module cleanup
We refactor the logic of the sitecustomize bootstrap script to
make it harder to introduce new initialisation logic that can
break some of the boilerplate bootstrapping logic, such as the
module cleanup mechanism for the support of frameworks like
gevent. We move all the logic into a dedicated module, preload,
and expose post_preload hooks in case that some code needs to be
run after the module cleanup. However, the problem is now slightly
shifted towards these hooks. Care must be taken not to register
any hooks that import modules that are not already imported during
the execution of the preload script.