fix(internal): ignore non-standard modules
When mapping imported modules to their origin, we handle
non-standard objects to ensure that we do not raise an exception
if we find a module that either has no origin, or that cannot be
referenced weakly. We expect these cases to be of no importance
to the tracking process, as these modules are unlikely to
provide useful information via the Python runtime.