chore(debugging): improve probe installation failure message
A probe might fail to install because the function it should be
installed into was not discovered. This could happen when the
function in question is decorated with a decorator that we do
not support. We improve the error message logging by hinting at
this possibility so that the user is better informed about the
reason why a probe failed to install.