feat(openai): add request duration for llmobs payloads (#7840)
This PR follows up on #7663 by adding the request duration metric of the
OpenAI Completion/ChatCompletion call to payloads being sent to the
LLMObs backend.

Because of the fact that the `openai.request` span has not been finished
before we can calculate the request duration, we calculate the request
duration roughly as the difference between the span start time and the
timestamp for when we start generating the llmobs record payload. Note
that this is not exactly accurate (off by approximately ~15-20ms for a
~1-2s request) compared to using the underlying `http.request` span. The
reason this approach was taken is that the `http.request` span is not
always guaranteed (requires the http requests integration to be
enabled).

### Next Steps

To allow a more accurate calculation of the request duration in the
future, we will need to compute the timestamp at the earliest possible
timestamp after receiving the response (in
`_EndpointHook.handle_request()`) and passing that timestamp all the way
to the `generate_llm_*completion()` function. This is not trivial which
is why we're choosing the simplest option in this PR for now.

Note: this PR does not need a changelog as LLMObs is still in private
beta and guarded by the feature flag `DD_OPENAI_LLMOBS_ENABLED`.

## Checklist

- [x] Change(s) are motivated and described in the PR description.
- [x] Testing strategy is described if automated tests are not included
in the PR.
- [x] Risk is outlined (performance impact, potential for breakage,
maintainability, etc).
- [x] Change is maintainable (easy to change, telemetry, documentation).
- [x] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
are followed. If no release note is required, add label
`changelog/no-changelog`.
- [x] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/)).
- [x] Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist

- [x] Title is accurate.
- [x] No unnecessary changes are introduced.
- [x] Description motivates each change.
- [x] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes unless absolutely necessary.
- [x] Testing strategy adequately addresses listed risk(s).
- [x] Change is maintainable (easy to change, telemetry, documentation).
- [x] Release note makes sense to a user of the library.
- [x] Reviewer has explicitly acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment.
- [x] Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)
- [x] If this PR touches code that signs or publishes builds or
packages, or handles credentials of any kind, I've requested a review
from `@DataDog/security-design-and-guidance`.
- [x] This PR doesn't touch any of that.