chore(tracer): report changes to config values
A feature is being built to report configuration in the Datadog UI based
on the telemetry data. In order to provide the most up to date
configuration data in the UI we need to report configuration change
events when settings are updated. The feature also supports reporting
the origin of a setting at the granularity of ("code", "env_var",
"remote_config" or "default") which will help customers debug where a
setting for their service might be coming from.

I noticed that the telemetry config changed code doesn't do any batching
of configuration updates so this would be a potential performance
improvement in the future.

Tests are added to cover the different origins.