refactor(debugging): use new probe status endpoint
We refactor the upload logic to redirect the probe status messages to
the new dedicated endpoint on the agent. The upload is done via a
multipart POST request with the event.json file carrying the same JSON
payload that was being uploaded along with log signals with in the
previous implementation.