chore(tracer): report changes to config values (#7909)
A feature is being built to report configuration in the Datadog UI based
on the telemetry data. In order to provide the most up to date
configuration data in the UI we need to report configuration change
events when settings are updated. The feature also supports reporting
the origin of a setting at the granularity of ("code", "env_var",
"remote_config" or "default") which will help customers debug where a
setting for their service might be coming from.

I noticed that the telemetry config changed code doesn't do any batching
of configuration updates so this would be a potential performance
improvement in the future. There is the risk that if a customer is
frequently configuring a setting that many telemetry events will be
generated which could lead to performance regressions. I imagine this
case is small since up until
https://github.com/DataDog/dd-trace-py/pull/7489 global config updates
to sample rate would not have any effect.

Tests are added to cover the different origins.


![image](https://github.com/DataDog/dd-trace-py/assets/6321485/8cbfbb1e-c2ad-4c1e-984b-2170822d2264)


The implementation should be consistent with
https://github.com/DataDog/system-tests/pull/1734.