refactor: add ancestor runtime ID
Because of the current implementation of the remote configuration
client, configuration requests are made by the common ancestor process.
This means that child processes are invisible to the RCM backend. Some
products require information about the runtime ID to provide a good UX.
We add the possibility of retrieving the runtime ID of the common
ancestor process (the one that will communicate with the RCM backend)
from any child so that its runtime ID can be correlated with the
ancestor, and used by backend services that rely on the enablement
signal coming from RCM requests.