fix(tracer): handle deleted remote configurations (#7933)
The data format changed with the recent refactor the the remoteconfig
client. A `None` value used to be passed to the handling method when a
configuration was deleted.

A test case is added to cover the regression.

No release note is required because the feature isn't released yet.

(cherry picked from commit 26ede5ce7cac28e21ad1b4e249960d92e48c81b1)