fix(internal): remove __class__ access interceptor
The __class__ attribute of the chained loader was intercepted to allow
the object to look like it has the type of the wrapped loader. This was
done to allow some tools, such as slotscheck to work as expected. We
remove this special handling as it seems this is no longer a problem.

Addresses #7950.