feat(ci_visibilty): introduce CI Visibility-specific log handler controlled by env var (#7955)
This introduces the `DD_CIVISIBILITY_LOG_LEVEL` env var, which adds a
log handler to the root logger that filters messages specifically from
the CI Visibility-related paths (`ddtrace.internal.ci_visibility` and
`ddtrace.ext.git`). The use of `DD_TRACE_DEBUG` preempts this behavior
and maintains the original debug behavior.

This is to support better logging for products like the `pytest` plugin,
where logging to `pytest`'s plugin is both intrusive, and results in
double-logging due to the stream handler on the `DDTrace` logger.

The handler is added to the root logger, instead of the `DDTrace`
logger, in order to preserve the latter's log levels (specifically for
the `FileHandler`).

This PR enables the behavior by default for `pytest`, and is considered
a feature (the extra logging is not considered a breaking change).

## Checklist

- [x] Change(s) are motivated and described in the PR description.
- [x] Testing strategy is described if automated tests are not included
in the PR.
- [x] Risk is outlined (performance impact, potential for breakage,
maintainability, etc).
- [x] Change is maintainable (easy to change, telemetry, documentation).
- [x] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
are followed. If no release note is required, add label
`changelog/no-changelog`.
- [x] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/)).
- [x] Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist

- [x] Title is accurate.
- [x] No unnecessary changes are introduced.
- [x] Description motivates each change.
- [x] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes unless absolutely necessary.
- [x] Testing strategy adequately addresses listed risk(s).
- [x] Change is maintainable (easy to change, telemetry, documentation).
- [x] Release note makes sense to a user of the library.
- [x] Reviewer has explicitly acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment.
- [x] Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)
- [x] If this PR touches code that signs or publishes builds or
packages, or handles credentials of any kind, I've requested a review
from `@DataDog/security-design-and-guidance`.
- [x] This PR doesn't touch any of that.