test(di): remove function probe after context exit
We fix a problem with the exploration tests due to function probe
getting removed when a snapshot is created. This happens to be too early
because of the new EMITTING status, which requires the probe to be
registered after the signal context is exited. This issue affects just
exploration tests, where an option is made available of removing probes
after they have been triggered to improve the performance of the test
runs.