chore(internal): symbol inference (#6920)
We introduce logic to infer the symbols exported by a module. This can
be used by any product that requires symbol information to provide
features such as expression auto-completion. The actual enablement of
the symbol inference and upload is controlled via a remote configuration
signal. For this feature to work as intended, RC is an essential
internal requirement.

For performance reasons, the symbol inference is performed starting from
module objects, after they have been imported. This way we skip a second
parsing step, and reuse the result of the original operation performed
by CPython itself when the module was loaded for the first time.

Because the main consumer of this feature is Dynamic Instrumentation,
the information that is retrieved from each module reflects the DI
capabilities of discovering "instrumentable" symbols at runtime. Some
finer details (such as nested functions/classes) might therefore be
missing at this stage.

There is also a special handling for scenarios where child processes are
spawned via fork to parallelise the workload. Because each child process
would perform the same symbol discovery, the backend would receive
multiple uploads of duplicate information. To avoid wasting resources
and bandwidth, we added logic to ensure that uploads happen from the
parent process, and _at most_ one forked child process. The risk of this
approach is that, if fork is used for anything else other than the
worker child model, the symbol picture reconstructed on the backend site
might be incomplete. However, we are betting on this scenario to be very
rare, compared to the more common forked child process model.

## Checklist

- [x] Change(s) are motivated and described in the PR description.
- [x] Testing strategy is described if automated tests are not included
in the PR.
- [x] Risk is outlined (performance impact, potential for breakage,
maintainability, etc).
- [x] Change is maintainable (easy to change, telemetry, documentation).
- [x] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
are followed. If no release note is required, add label
`changelog/no-changelog`.
- [x] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/)).
- [x] Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist

- [ ] Title is accurate.
- [ ] No unnecessary changes are introduced.
- [ ] Description motivates each change.
- [ ] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes unless absolutely necessary.
- [ ] Testing strategy adequately addresses listed risk(s).
- [ ] Change is maintainable (easy to change, telemetry, documentation).
- [ ] Release note makes sense to a user of the library.
- [ ] Reviewer has explicitly acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment.
- [ ] Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)
- [ ] If this PR touches code that signs or publishes builds or
packages, or handles credentials of any kind, I've requested a review
from `@DataDog/security-design-and-guidance`.
- [ ] This PR doesn't touch any of that.