refactor(profiling): variable name if arguments available
We retrieve the variable name corresponding to the first argument only
if the code object actually has any arguments. This prevents us blindly
trying to retrieve the first entry of the variable name array in the
hope that it is one between `self` and `cls`. The benefit of this is a
less noisy debug log where all the failed attempts to resolve a local
variable that might not be set already are currently logged.