ci(suitespec): finer-grained internal component
The ddtrace/internal area of the code-base contains parts that are not
necessarily core functionality. For example, changes in symbol_db do not
need to trigger tests in the suite of the tracer. We spell out
explicitly what sources are part of the core library, and move the rest
to the more appropriate components. We also ensure that the new
components are covered by at least one appropriate test suite so that
changes to the files matched by the patters are actually covered in
tests.