feat(llmobs): set parent ID as the nearest LLMObs span in ancestor tree (#8431)
This PR changes generated LLMObs span events to ensure that a given
LLMObs span event has a parent ID which belongs to the nearest LLMObs
span in its ancestor tree rather than always its parent span.

This is because LLMObs doesn't store non-LLMObs span events, meaning
that in the case of traces with non-LLMObs spans, i.e. `LLMObs-span -->
APM span --> LLMObs-span` the last span's parent ID will be that of the
second (APM) span which isn't sent to LLMObs, meaning there will be a
gap/error when we try to recreate the trace. The solution is to only use
LLMObs spans to set the parent ID, i.e. the nearest span in the ancestor
tree which is also of type `LLM`.

This PR also adds tests for the `LLMObsTraceProcessor` and refactors
some commonly used fixtures/utils.

### Limitations and Future Steps
In cases of distributed traces, multithreading, or forked processes, a
span will not have access to all of the spans in its ancestor tree
beyond its immediate parent span (usually through a context). This means
that we will likely submit multiple trace fragments to LLMObs each with
its own "root" span that doesn't have a LLMObs parent ID. This is fine
for now as LLMObs does not currently support distributed tracing, but
for when we want to, we'll need to rethink how we're submitting LLMObs
spans.

The alternative approaches are:
- Submit all spans in a trace which contains a LLMObs span, which means
we won't have to find the nearest LLMObs span in a span's parent tree
(which will always be its immediate parent). Requires changes on LLMObs
backend/intake to accept non-LLM type spans.
- Propagate LLMObs parent ID to all spans, as well as on distributed
request/response headers. Not ideal as we will have to add tags to
potentially every span.


## Checklist

- [x] Change(s) are motivated and described in the PR description
- [x] Testing strategy is described if automated tests are not included
in the PR
- [x] Risks are described (performance impact, potential for breakage,
maintainability)
- [x] Change is maintainable (easy to change, telemetry, documentation)
- [x] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
are followed or label `changelog/no-changelog` is set
- [x] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/))
- [x] Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))
- [x] If this PR changes the public interface, I've notified
`@DataDog/apm-tees`.
- [x] If change touches code that signs or publishes builds or packages,
or handles credentials of any kind, I've requested a review from
`@DataDog/security-design-and-guidance`.

## Reviewer Checklist

- [x] Title is accurate
- [x] All changes are related to the pull request's stated goal
- [x] Description motivates each change
- [x] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- [x] Testing strategy adequately addresses listed risks
- [x] Change is maintainable (easy to change, telemetry, documentation)
- [x] Release note makes sense to a user of the library
- [x] Author has acknowledged and discussed the performance implications
of this PR as reported in the benchmarks PR comment
- [x] Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)