chore(rc): add skip shutdown option
We add an internal option to RC to skip the shutdown step and prevent
products from being unregistered. This is a workaroud for adding support
for systems such as Django Q. On platforms like Linux, these systems
use multiprocessing, and the default fork start method, to spawn child
processes. The main process then invokes the shutdown, potentially
leaving the child processes without an RC client to listen to for new
configuration payload. With the proposed changes, we prevent the RC
client in the parent process to responding to the shutdown, thus
keeping the RC poller worker alive. We also prevent products from
being unregistered, which would also prevent configuration from being
fetched for these products.