deprecations: deprecate `[]` notation for `DD_TRACE_METHODS` (#8438)
## Description
This PR deprecates the `[]` notation for `DD_TRACE_METHODS` in favor of
the `:` notation.

When customers tried to add `django_q` methods to `DD_TRACE_METHODS`,
calling `__import__` during bootstrapping to figure out whether the
provided method was a module or class method caused boostrapping errors.
This is because modules like `django_q` attempt to do configurations
upon attempting to import, which we don't want happening during
bootstrapping.

The notation change for `DD_TRACE_METHODS` is primarily to help us
distinguish between the legacy/deprecated and new semantics for
`DD_TRACE_METHODS`. With the deprecated semantics, we needed to guess
what the base module was by importing, since only the method name is
within the brackets even if it's actually a class method. With the new
semantics, we are asking customers to include the class with the method
name if it's a class method.

Example:
```
DD_TRACE_METHODS="mymodule.mysubmodule.myclass[myfunc,otherfunc];..."
```
Here we would attempt to `__import__(mymodule.mysubmodule.myclass)`,
fail because the class is included, then re-attempt with
`__import__(mymodule.mysubmodule)`, which would succeed and indicate the
base module is only `mymodule.mysubmodule`, and the functions are a
class function called `myclass.myfunc` and `myclass.otherfunc`.

With the new semantics, this guess work would be taken out.
```
DD_TRACE_METHODS="mymodule.mysubmodule:myclass.myfunc,myclass.otherfunc;..."
```
Here we know for sure the base module is just `mymodule.mysubmodule`,
and the function names are also explicitly listed.

## Risks
This is a customer facing change, so respective updates to customer
facing documentation are included. After the next major release,
customers will be required to switch to the updated semantics.

## Checklist

- [x] Change(s) are motivated and described in the PR description
- [x] Testing strategy is described if automated tests are not included
in the PR
- [x] Risks are described (performance impact, potential for breakage,
maintainability)
- [x] Change is maintainable (easy to change, telemetry, documentation)
- [x] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
are followed or label `changelog/no-changelog` is set
- [x] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/))
- [x] Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))
- [ ] If this PR changes the public interface, I've notified
`@DataDog/apm-tees`.
- [x] If change touches code that signs or publishes builds or packages,
or handles credentials of any kind, I've requested a review from
`@DataDog/security-design-and-guidance`.

## Reviewer Checklist

- [ ] Title is accurate
- [ ] All changes are related to the pull request's stated goal
- [ ] Description motivates each change
- [ ] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- [ ] Testing strategy adequately addresses listed risks
- [ ] Change is maintainable (easy to change, telemetry, documentation)
- [ ] Release note makes sense to a user of the library
- [ ] Author has acknowledged and discussed the performance implications
of this PR as reported in the benchmarks PR comment
- [ ] Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)


[APMS-10881]:
https://datadoghq.atlassian.net/browse/APMS-10881?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ